[ccassandra/extract.js] Try fix from tardypad#89 (#31)
Even though this fix was originally dismissed as not elegant, it is far better than a broken CCassandra plugin!  In general, I disagree with the notion that not everything reasonably feasible with JavaScript is allowed.  For details, see https://github.com/tardypad/sailfishos-daily-comics/issues/89

Furthermore this scheme seems to be generically applicable to comic strips from `tapas.io`.